Reorganized and reformatted original files.
